Rename ResourceRoutingRepository. It handles more than just routing nowadays.
