+++ src/Resource.php
@@ -17,7 +17,6 @@ use Statamic\Statamic;
 use StatamicRadPack\Runway\Exceptions\PublishedColumnMissingException;
 use StatamicRadPack\Runway\Fieldtypes\BelongsToFieldtype;
 use StatamicRadPack\Runway\Fieldtypes\HasManyFieldtype;
-use StatamicRadPack\Runway\Routing\ResourceRoutingRepository;
 
 class Resource
 {
@@ -43,7 +42,7 @@ class Resource
         return $this->model->newQuery()
             ->runway()
             ->afterQuery(function ($models) {
-                return app(ResourceRoutingRepository::class)->applySubstitutions($models);
+                return app(ModelRepository::class)->applySubstitutions($models);
             });
     }
