+++ src/Traits/HasRunwayResource.php
@@ -20,7 +20,6 @@ use StatamicRadPack\Runway\Data\HasAugmentedInstance;
 use StatamicRadPack\Runway\Fieldtypes\HasManyFieldtype;
 use StatamicRadPack\Runway\Relationships;
 use StatamicRadPack\Runway\Resource;
-use StatamicRadPack\Runway\Routing\ResourceRoutingRepository;
 use StatamicRadPack\Runway\Runway;
 
 trait HasRunwayResource
